ScoreAPI/requirements.txt
fastapi
uvicorn
boto3
python-decouple
httpx
python-multipart
